backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary,pool]
pydantic[email]
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
python-slugify
